+++ src/main/java/com/example/fundbox24backend/api/model/LostReport.java
@@ -20,11 +20,11 @@ public class LostReport extends Report
     @NotNull
     @OneToOne private Location lastSeenLocation;
     @OneToOne private Location lostLocation = lastSeenLocation;
-    private int lostRadius;
+    private double lostRadius;
     @NotNull
     @ManyToOne private User owner;
 
-    public LostReport(String title, String description, String imagePath, boolean isFinished, Category category, LocalDateTime lastSeenDate, Location lastSeenLocation, Location lostLocation, int lostRadius, User owner)
+    public LostReport(String title, String description, String imagePath, boolean isFinished, Category category, LocalDateTime lastSeenDate, Location lastSeenLocation, Location lostLocation, double lostRadius, User owner)
     {
         super(title, description, imagePath, isFinished, category);
         this.lastSeenDate = lastSeenDate;
